[CON-4908] Send SN notification on plugin disable
